+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0
+
+- **BREAKING:** `PausableTimer` is now `final`
+- Minor cleanups and fixes in documentations and tests
+
 ## 2.0.0+1
 
 - Update and test against Dart 3.1.